# src/server/docs/swagger.yaml
basePath: /api/v1
definitions:
  auth.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  auth.LoginResponse:
    properties:
      token:
        type: string
      username:
        type: string
    type: object
  deployment.FunctionListResponse:
    properties:
      name:
        type: string
      runtime:
        type: string
    type: object
  deployment.FunctionResponse:
    properties:
      created_at:
        type: string
      handler:
        type: string
      id:
        type: string
      memory:
        type: integer
      module:
        type: string
      name:
        type: string
      runtime:
        type: string
      timeout:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is the API for the Oblak Serverless platform.
  title: Oblak Serverless API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user using credentials and returns an API token.
      parameters:
      - description: User credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successful login
          schema:
            $ref: '#/definitions/auth.LoginResponse'
        "400":
          description: Invalid request payload
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Invalid credentials
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: Rate limit exceeded
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User login
      tags:
      - auth
  /functions/:
    get:
      description: Retrieves a list of all deployed functions belonging to the authenticated
        user.
      produces:
      - application/json
      responses:
        "200":
          description: List of functions
          schema:
            items:
              $ref: '#/definitions/deployment.FunctionListResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: Rate limit exceeded
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List functions
      tags:
      - functions
  /functions/{name}:
    delete:
      description: Deletes a function from the database and removes its artifact from
        the storage.
      parameters:
      - description: Function name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Function not found
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: Rate limit exceeded
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to delete artifact or database record
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a function
      tags:
      - functions
    get:
      description: Retrieves detailed information about a specific function by its
        name.
      parameters:
      - description: Function name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Function details
          schema:
            $ref: '#/definitions/deployment.FunctionResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Function not found
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: Rate limit exceeded
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get function details
      tags:
      - functions
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"